Renamed PqEventManager to PqEvents
